coolshortflask/requirements.txt
click==8.1.3
docker==6.0.0
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
markupsafe==2.1.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
setuptools==60.2.0
Werkzeug==2.2.2
wheel==0.37.1
zipp==3.9.0
